+++ ai_spider/fine_tune.py
@@ -131,8 +131,8 @@ async def fine_tune_task(request, body, job_id, user_id):
     gpu_filter["capabilities"] = ["llama-fine-tune"]
     job = fine_tuning_jobs_db[user_id][job_id]
     mbase = body.model.split("/")[-1]
-    lora_key = f"{user_id}/{mbase}:{job_id}.lora.gz"
-    gguf_key = f"{user_id}/{mbase}:{job_id}.gguf"
+    lora_key = f"{user_id}/{mbase}.{job_id}.lora.gz"
+    gguf_key = f"{user_id}/{mbase}.{job_id}.gguf"
     s3 = await get_s3()
     upload = {
         "lora": {
